- initial commit for DiSTNet2D (in process) - minor changes in GUI

+++ bacmman-core/src/main/java/bacmman/processing/neighborhood/NeighborhoodFactory.java
@@ -25,29 +25,25 @@ package bacmman.processing.neighborhood;
 public class NeighborhoodFactory {
     
 
-    public static int[][] getNeighbourhood3D(float radius, float radiusZ, float thickness) {
+    public static int[][] getNeighbourhood3D(double radius, double radiusZ, double thickness) {
         if (radiusZ > 0 && radiusZ < 1) {
             radiusZ = 1;
         }
-        //IJ.log("neigh: XY:"+radius+" Z:"+radiusZ+ " Thickness:"+thickness);
-        float r = (float) radius / radiusZ;
+        double r = (float) radius / radiusZ;
         int rad = (int) (radius + 0.5f);
         int radZ = (int) (radiusZ + 0.5f);
         int[][] temp = new int[3][(2 * rad + 1) * (2 * rad + 1) * (2 * radZ + 1)];
-        //float[] tempDist = new float[temp[0].length];
         int count = 0;
-        float rad2 = radius * radius;
-        float radMin = (radius >= thickness) ? (float) Math.pow(radius - thickness, 2) : Float.MIN_VALUE; //:exclusion du point central
+        double rad2 = radius * radius;
+        double radMin = (radius >= thickness) ?  Math.pow(radius - thickness, 2) : Double.MIN_VALUE; //:exclusion du point central
         for (int zz = -radZ; zz <= radZ; zz++) {
             for (int yy = -rad; yy <= rad; yy++) {
                 for (int xx = -rad; xx <= rad; xx++) {
-                    float d2 = zz * r * zz * r + yy * yy + xx * xx;
+                    double d2 = zz * r * zz * r + yy * yy + xx * xx;
                     if (d2 <= rad2 && d2 > radMin && !((xx == 0) && (yy == 0) && (zz == 0))) {
                         temp[0][count] = xx;
                         temp[1][count] = yy;
                         temp[2][count] = zz;
-                        //tempDist[count] = (float) Math.sqrt(d2);
-                        //IJ.log("neigh: X:"+xx+" Y:"+yy+ " Z:"+zz);
                         count++;
                     }
                 }
@@ -57,33 +53,30 @@ public class NeighborhoodFactory {
         System.arraycopy(temp[0], 0, res[0], 0, count);
         System.arraycopy(temp[1], 0, res[1], 0, count);
         System.arraycopy(temp[2], 0, res[2], 0, count);
-        //ij.IJ.log("Neigbor: Radius:"+radius+ " radiusZ"+radiusZ+ " count:"+count);
         return res;
     }
 
-    public static int[][] getHalfNeighbourhood(float radius, float radiusZ, float thickness) {
+    public static int[][] getHalfNeighbourhood(double radius, double radiusZ, double thickness) {
         if (radiusZ > 0 && radiusZ < 1) {
             radiusZ = 1;
         }
         //IJ.log("neigh: XY:"+radius+" Z:"+radiusZ+ " Thickness:"+thickness);
-        float r = (float) radius / radiusZ;
+        double r = (float) radius / radiusZ;
         int rad = (int) (radius + 0.5f);
         int radZ = (int) (radiusZ + 0.5f);
         int[][] temp = new int[3][(2 * rad + 1) * (2 * rad + 1) * (2 * radZ + 1)];
         //float[] tempDist = new float[temp[0].length];
         int count = 0;
-        float rad2 = radius * radius;
-        float radMin = (radius >= thickness) ? (float) Math.pow(radius - thickness, 2) : Float.MIN_VALUE; //exclusion du point central
+        double rad2 = radius * radius;
+        double radMin = (radius >= thickness) ? Math.pow(radius - thickness, 2) : Double.MIN_VALUE; //exclusion du point central
         for (int zz = 0; zz <= radZ; zz++) {
             for (int yy = (zz == 0) ? 0 : -rad; yy <= rad; yy++) {
                 for (int xx = (zz == 0 && yy == 0) ? -rad + 1 : -rad; xx <= rad; xx++) {
-                    float d2 = zz * r * zz * r + yy * yy + xx * xx;
+                    double d2 = zz * r * zz * r + yy * yy + xx * xx;
                     if (d2 <= rad2 && d2 > radMin) {
                         temp[0][count] = xx;
                         temp[1][count] = yy;
                         temp[2][count] = zz;
-                        //tempDist[count] = (float) Math.sqrt(d2);
-                        //IJ.log("neigh: X:"+xx+" Y:"+yy+ " Z:"+zz);
                         count++;
                     }
                 }
@@ -93,7 +86,6 @@ public class NeighborhoodFactory {
         System.arraycopy(temp[0], 0, res[0], 0, count);
         System.arraycopy(temp[1], 0, res[1], 0, count);
         System.arraycopy(temp[2], 0, res[2], 0, count);
-        //ij.IJ.log("Neigbor: Radius:"+radius+ " radiusZ"+radiusZ+ " count:"+count);
         return res;
     }
